+++ src/main/java/io/mapsmessaging/consul/ecwid/EcwidConsulManager.java
@@ -18,6 +18,9 @@
 package io.mapsmessaging.consul.ecwid;
 
 import io.mapsmessaging.consul.ConsulServerApi;
+import io.mapsmessaging.logging.Logger;
+import io.mapsmessaging.logging.LoggerFactory;
+import io.mapsmessaging.logging.ServerLogMessages;
 import io.mapsmessaging.network.io.EndPointServer;
 import io.mapsmessaging.rest.RestApiServerManager;
 
@@ -25,14 +28,33 @@ import java.io.IOException;
 import java.util.List;
 import java.util.Map;
 
+import static io.mapsmessaging.logging.ServerLogMessages.CONSUL_CLIENT_LOG;
+
 public class EcwidConsulManager extends ConsulServerApi {
 
-  public EcwidConsulManager(String name) {
+  private final Logger logger = LoggerFactory.getLogger(EcwidConsulManager.class);
+
+  public EcwidConsulManager(String name) throws IOException {
     super(name);
+    try {
+      logger.log(CONSUL_CLIENT_LOG, "Creating client", consulConfiguration);
+      //    client = createBuilder().build();
+      logger.log(CONSUL_CLIENT_LOG, "Created client", consulConfiguration);
+      //  agentClient = consulConfiguration.registerAgent() ? client.agentClient() : null;
+      //  keyValueClient = client.keyValueClient();
+      logger.log(ServerLogMessages.CONSUL_STARTUP);
+    } catch (Exception ex) {
+      throw new IOException(ex);
+    }
   }
 
   @Override
   public void stop() {
+    super.stop();
+  }
+
+  @Override
+  protected void pingService() {
 
   }
